Extract retrySubmitMixin
